edge/requirements.txt
pyyaml
requests
websockets
numpy
opencv-python-headless
onnxruntime
pillow
